+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Added struct member reference and completion support
 - Added structure view feature
 - Added breadcrumbs support
+- Added structure-aware navigation bar feature
 - Minor feature improvements
 
 ## 0.5.0-beta (2026-01-22)